+++ src/main/java/dev/chenjr/attendance/service/IUserService.java
@@ -46,6 +46,6 @@ public interface IUserService extends IService {
     boolean userExists(long uid);
 
     boolean userExists(String account);
-
-    void deleteByUid(Integer uid);
+    
+    void deleteByUid(long uid);
 }
